# docs/swagger.yaml
basePath: /v1
definitions:
  handlers.AuthToken:
    properties:
      token:
        type: string
    type: object
  handlers.DoctorLogin:
    properties:
      Email:
        type: string
      Password:
        type: string
    type: object
  handlers.DoctorRegistration:
    properties:
      AadharNumber:
        type: string
      Address:
        type: string
      BloodGroup:
        type: string
      ClinicName:
        type: string
      DOB:
        type: string
      Email:
        type: string
      Experience:
        type: string
      FirstName:
        type: string
      Gender:
        type: string
      LastName:
        type: string
      LicenseNumber:
        type: string
      MobileNumber:
        type: string
      Password:
        type: string
      Specialty:
        type: string
    type: object
  handlers.Message:
    properties:
      message:
        type: string
    type: object
  handlers.PatientLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.PatientRegistration:
    properties:
      AadharNumber:
        type: string
      Address:
        type: string
      BloodGroup:
        type: string
      DOB:
        type: string
      Email:
        type: string
      EmergencyContact:
        type: string
      FirstName:
        type: string
      Gender:
        type: string
      LastName:
        type: string
      MobileNumber:
        type: string
      Password:
        type: string
    type: object
  handlers.PrescriptionCreated:
    properties:
      message:
        type: string
      prescription:
        $ref: '#/definitions/models.Prescription'
    type: object
  handlers.PrescriptionInput:
    properties:
      diagnosis:
        type: string
      doctorId:
        type: integer
      followUpDate:
        type: string
      instructions:
        type: string
      medications:
        type: string
      patientId:
        type: integer
    type: object
  handlers.VisitInput:
    properties:
      doctor_id:
        type: integer
      notes:
        type: string
      patient_id:
        type: integer
      visit_date:
        type: string
    type: object
  models.Prescription:
    properties:
      createdAt:
        type: string
      diagnosis:
        type: string
      doctorId:
        type: integer
      followUpDate:
        type: string
      id:
        type: integer
      instructions:
        type: string
      medications:
        type: string
      patientId:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Visit:
    properties:
      created_at:
        type: string
      doctor_id:
        type: integer
      id:
        type: integer
      notes:
        type: string
      patient_id:
        type: integer
      updated_at:
        type: string
      visit_date:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API documentation for the Medicare service.
  title: Medicare API
  version: 1.0.0
paths:
  /doctor/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Doctor login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DoctorLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthToken'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login as a doctor
      tags:
      - doctor
  /doctor/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Doctor registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DoctorRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthToken'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new doctor
      tags:
      - doctor
  /patient/auth/login:
    post:
      consumes:
      - application/json
      description: Also handles /patient/auth/logout (currently routed to the same
        handler).
      parameters:
      - description: Patient login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatientLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthToken'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login as a patient
      tags:
      - patient
  /patient/auth/logout:
    post:
      consumes:
      - application/json
      description: Also handles /patient/auth/logout (currently routed to the same
        handler).
      parameters:
      - description: Patient login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatientLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthToken'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login as a patient
      tags:
      - patient
  /patient/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Patient registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatientRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthToken'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new patient
      tags:
      - patient
  /prescriptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Prescription
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PrescriptionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PrescriptionCreated'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new prescription
      tags:
      - prescriptions
  /prescriptions/{id}:
    delete:
      parameters:
      - description: Prescription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Message'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a prescription
      tags:
      - prescriptions
    get:
      parameters:
      - description: Prescription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Prescription'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a prescription by ID
      tags:
      - prescriptions
  /prescriptions/patient/{patientID}:
    get:
      parameters:
      - description: Patient ID
        in: path
        name: patientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Prescription'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List prescriptions by patient
      tags:
      - prescriptions
  /visits:
    post:
      consumes:
      - application/json
      parameters:
      - description: Visit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VisitInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new visit
      tags:
      - visits
  /visits/{id}:
    delete:
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a visit
      tags:
      - visits
    get:
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Visit'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a visit by ID
      tags:
      - visits
    put:
      consumes:
      - application/json
      parameters:
      - description: Visit ID
        in: path
        name: id
        required: true
        type: integer
      - description: Visit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VisitInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update a visit
      tags:
      - visits
  /visits/doctor/{doctorID}:
    get:
      parameters:
      - description: Doctor ID
        in: path
        name: doctorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Visit'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get visits by doctor ID
      tags:
      - visits
  /visits/patient/{patientID}:
    get:
      parameters:
      - description: Patient ID
        in: path
        name: patientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Visit'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get visits by patient ID
      tags:
      - visits
swagger: "2.0"